requirements.txt
bs4
pyyaml
pysftp
requests
guessit
feedparser